# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * RocketFi API — Token Launchpad Backend
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Token {
  id: number;
  address: string;
  name: string;
  symbol: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  creatorAddress: string;
  /** BigInt as string */
  virtualTokenReserves: string;
  /** BigInt as string */
  virtualEthReserves: string;
  /** BigInt as string */
  realTokenReserves: string;
  /** BigInt as string */
  realEthReserves: string;
  /** BigInt as string */
  totalSupply: string;
  /** @nullable */
  marketCapEth?: string | null;
  /** @nullable */
  priceEth?: string | null;
  graduated: boolean;
  volumeEth: string;
  tradeCount: number;
  holderCount: number;
  /** @nullable */
  twitterUrl?: string | null;
  /** @nullable */
  telegramUrl?: string | null;
  /** @nullable */
  websiteUrl?: string | null;
  createdAt: string;
}

export interface TokenInput {
  address: string;
  name: string;
  symbol: string;
  description?: string;
  imageUrl?: string;
  creatorAddress: string;
  totalSupply: string;
  virtualTokenReserves: string;
  virtualEthReserves: string;
  twitterUrl?: string;
  telegramUrl?: string;
  websiteUrl?: string;
}

export interface TokenUpdate {
  virtualTokenReserves?: string;
  virtualEthReserves?: string;
  realTokenReserves?: string;
  realEthReserves?: string;
  marketCapEth?: string;
  priceEth?: string;
  graduated?: boolean;
  volumeEth?: string;
  tradeCount?: number;
  holderCount?: number;
}

export interface Trade {
  id: number;
  tokenAddress: string;
  /** @nullable */
  tokenName?: string | null;
  /** @nullable */
  tokenSymbol?: string | null;
  traderAddress: string;
  isBuy: boolean;
  ethAmount: string;
  tokenAmount: string;
  /** @nullable */
  priceEth?: string | null;
  txHash: string;
  timestamp: string;
}

export interface TradeInput {
  traderAddress: string;
  isBuy: boolean;
  ethAmount: string;
  tokenAmount: string;
  priceEth?: string;
  txHash: string;
  timestamp: string;
}

export interface PlatformStats {
  totalTokens: number;
  totalGraduated: number;
  totalVolumeEth: string;
  totalTrades: number;
  tokensLast24h: number;
  volumeLast24h: string;
  tradesLast24h: number;
}

export interface Profile {
  address: string;
  username: string;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  twitterHandle?: string | null;
  /** @nullable */
  websiteUrl?: string | null;
  followersCount: number;
  followingCount: number;
  createdAt: string;
  updatedAt: string;
}

export interface ProfileInput {
  address: string;
  username?: string;
  bio?: string;
  avatarUrl?: string;
  twitterHandle?: string;
  websiteUrl?: string;
}

export interface ProfileUpdate {
  username?: string;
  bio?: string;
  avatarUrl?: string;
  twitterHandle?: string;
  websiteUrl?: string;
}

export interface ActivityItem {
  id: number;
  tokenAddress: string;
  tokenName: string;
  tokenSymbol: string;
  /** @nullable */
  tokenImageUrl?: string | null;
  traderAddress: string;
  isBuy: boolean;
  ethAmount: string;
  tokenAmount: string;
  txHash: string;
  timestamp: string;
}

export type ListTokensParams = {
sort?: ListTokensSort;
limit?: number;
offset?: number;
search?: string;
graduated?: boolean;
};

export type ListTokensSort = typeof ListTokensSort[keyof typeof ListTokensSort];


export const ListTokensSort = {
  newest: 'newest',
  trending: 'trending',
  volume: 'volume',
  marketcap: 'marketcap',
} as const;

export type GetTrendingTokensParams = {
limit?: number;
};

export type GetRecentActivityParams = {
limit?: number;
};