# lib/api-zod/src/generated/types/activityItem.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * RocketFi API — Token Launchpad Backend
 * OpenAPI spec version: 0.1.0
 */

export interface ActivityItem {
  id: number;
  tokenAddress: string;
  tokenName: string;
  tokenSymbol: string;
  /** @nullable */
  tokenImageUrl?: string | null;
  traderAddress: string;
  isBuy: boolean;
  ethAmount: string;
  tokenAmount: string;
  txHash: string;
  timestamp: Date;
}