# lib/api-zod/src/generated/types/platformStats.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * RocketFi API — Token Launchpad Backend
 * OpenAPI spec version: 0.1.0
 */

export interface PlatformStats {
  totalTokens: number;
  totalGraduated: number;
  totalVolumeEth: string;
  totalTrades: number;
  tokensLast24h: number;
  volumeLast24h: string;
  tradesLast24h: number;
}