# lib/api-zod/src/generated/types/profile.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * RocketFi API — Token Launchpad Backend
 * OpenAPI spec version: 0.1.0
 */

export interface Profile {
  address: string;
  username: string;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  twitterHandle?: string | null;
  /** @nullable */
  websiteUrl?: string | null;
  followersCount: number;
  followingCount: number;
  createdAt: Date;
  updatedAt: Date;
}