# lib/api-zod/src/generated/types/token.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * RocketFi API — Token Launchpad Backend
 * OpenAPI spec version: 0.1.0
 */

export interface Token {
  id: number;
  address: string;
  name: string;
  symbol: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  creatorAddress: string;
  /** BigInt as string */
  virtualTokenReserves: string;
  /** BigInt as string */
  virtualEthReserves: string;
  /** BigInt as string */
  realTokenReserves: string;
  /** BigInt as string */
  realEthReserves: string;
  /** BigInt as string */
  totalSupply: string;
  /** @nullable */
  marketCapEth?: string | null;
  /** @nullable */
  priceEth?: string | null;
  graduated: boolean;
  volumeEth: string;
  tradeCount: number;
  holderCount: number;
  /** @nullable */
  twitterUrl?: string | null;
  /** @nullable */
  telegramUrl?: string | null;
  /** @nullable */
  websiteUrl?: string | null;
  createdAt: Date;
}