# lib/api-zod/src/generated/types/tokenInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * RocketFi API — Token Launchpad Backend
 * OpenAPI spec version: 0.1.0
 */

export interface TokenInput {
  address: string;
  name: string;
  symbol: string;
  description?: string;
  imageUrl?: string;
  creatorAddress: string;
  totalSupply: string;
  virtualTokenReserves: string;
  virtualEthReserves: string;
  twitterUrl?: string;
  telegramUrl?: string;
  websiteUrl?: string;
}