# lib/api-zod/src/generated/types/tokenUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * RocketFi API — Token Launchpad Backend
 * OpenAPI spec version: 0.1.0
 */

export interface TokenUpdate {
  virtualTokenReserves?: string;
  virtualEthReserves?: string;
  realTokenReserves?: string;
  realEthReserves?: string;
  marketCapEth?: string;
  priceEth?: string;
  graduated?: boolean;
  volumeEth?: string;
  tradeCount?: number;
  holderCount?: number;
}