# lib/api-zod/src/generated/types/trade.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * RocketFi API — Token Launchpad Backend
 * OpenAPI spec version: 0.1.0
 */

export interface Trade {
  id: number;
  tokenAddress: string;
  /** @nullable */
  tokenName?: string | null;
  /** @nullable */
  tokenSymbol?: string | null;
  traderAddress: string;
  isBuy: boolean;
  ethAmount: string;
  tokenAmount: string;
  /** @nullable */
  priceEth?: string | null;
  txHash: string;
  timestamp: Date;
}